week 4 beginner b begun

+++ Week 1 Beginner A.cpp
@@ -4,11 +4,11 @@
 #include <cmath>
 
 int num_souvenirs;
-int budget; 
-std::vector<long> souvenir_base_prices;
+long budget; 
+std::vector<unsigned long> souvenir_base_prices;
 
-int calculate_total_price(int tentative_num_souvenirs) {
-    std::vector<long> souvenir_prices = {};
+long calculate_total_price(int tentative_num_souvenirs) {
+    std::vector<unsigned long> souvenir_prices = {};
     for (int i = 0; i < num_souvenirs; i++) {
         souvenir_prices.push_back(souvenir_base_prices[i] + (i + 1) * tentative_num_souvenirs);
     }
@@ -28,7 +28,7 @@ void final_solution() {
     int num_souvenirs_upper_bound = num_souvenirs;          // inclusive
     int num_souvenirs_lower_bound = 0;                      // inclusive
     int tentative_num_souvenirs;
-    int tentative_total_price;
+    long tentative_total_price;
     while (num_souvenirs_upper_bound != num_souvenirs_lower_bound) {
         tentative_num_souvenirs = ceil((num_souvenirs_upper_bound + num_souvenirs_lower_bound) / 2.0);
         tentative_total_price = calculate_total_price(tentative_num_souvenirs);
